refactor: Se deja global a la apliación el método de obtener el id de la cooperativa.
- Se cambia donde estaba fijo en "1" el id de coperativa por el método global,
  en lo relacionado en las funcionalidades de cliente.
