+++ Cliente/utils.py
@@ -1,11 +1,12 @@
 from MarketPlace.models import Catalogo_Producto
-from MarketPlace.utils import cantidad_disponible_producto_catalogo, get_or_create_week
+from MarketPlace.utils import cantidad_disponible_producto_catalogo, get_or_create_week, get_id_cooperativa_global
 
 
 def agregar_producto_carrito(request, product_id, quantity):
-    product = Catalogo_Producto.objects.get(fk_producto_id=product_id,fk_catalogo__fk_cooperativa_id = 1,
+    product = Catalogo_Producto.objects.get(fk_producto_id=product_id,
+                                            fk_catalogo__fk_cooperativa_id=get_id_cooperativa_global(request),
                                             fk_catalogo__fk_semana=get_or_create_week())
-    cantidad_disponible = cantidad_disponible_producto_catalogo(product, 1)
+    cantidad_disponible = cantidad_disponible_producto_catalogo(product, get_id_cooperativa_global(request))
     cart = request.session.get('cart', None)
     if not cart:
         cart = {
